requirements.txt
requests>=2.28.0
httpx[http2,brotli]>=0.24.0
httpcore[asyncio]>=0.17.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
lxml>=4.9.0
urllib3>=1.26.0
PySide6>=6.4.0
certifi>=2023.0.0
mitmproxy>=10.0.0
